Move logging to RejectionDataFactories
