+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/FunctionHost/Common/ValidationErrorLogMessageBuilderTests.cs
@@ -18,7 +18,7 @@ using FluentAssertions;
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
 using GreenEnergyHub.Charges.Domain.Dtos.Validation;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
-using GreenEnergyHub.Charges.FunctionHost.Common;
+using GreenEnergyHub.Charges.MessageHub.Models.Shared;
 using GreenEnergyHub.Charges.Tests.Builders.Command;
 using GreenEnergyHub.TestHelpers;
 using Xunit;
